In the Compile_Code_Controller, the compile_code() function, we fixed a bug involving the spacing inside the file path. We added quotation marks and the start of ends of the file path to prevent this error.
In the error.h, we added a new error for no file name specified, and included this check in the File_Reader, the open_file() function and in Open_Existing_File_Controller, the open_existing_file() function.

In the Open_Existing_File_Controller, the open_existing_file() function, we added an additional check for the extension of an opened file, if it isnt .c or .cpp, we will not load the file, and will popup a window with a corresponding warning.