Included the keystone API to work with both QT Creator and Visual Studios 2017.
Added a morph button and a morph menu item into the UI.

Modified windows.h into pe32.h, to only use relevant structs for reading 32 bit pe files.

In the Compile_Code_Controller, we added a new variable, exe_file_path, along with its relevant accessors and mutators

Added the Morph_Executable_Controller, in BCE format, this will allow us to morph the executable file by reading via a new method, to read bytes, in the File_Reader. Currently, we have validated the DOS and PE signatures and increase the number of sections. In addition, we have made a function to generate a random 8 character name for the new payload section. Note that as of yet, no morphing or writing to file has been implemented.

Added new relevant errors and success messages into error.h

We have also noted that printf() behaves abnormally with the QT application. It will only display if there is a std::cout statement after it.