In the Morph_Executable_Controller, we added the capstone api, which will convert machine code to assembly instructions, this is needed for calculation when adding alternative and junk instructions.
We have created a container class, called Disassembly, which stores the data from the capstone api. This class is used when  adding alternative and junk instructions.

In the Morph_Executable_Controller, in the adding junk code function, junk code will be added randomly to the instructions of the Disassembly class vector. This adding of junk code will factor in the relative jump offsets when adding the code, hence, the relative jump offsets will be adjusted accordingly. Several methods were created to perform the aforementioned.

In the Morph_Executable_Controller, there are several functions to replace certain instructions with an alternate equivalent. These following instructions : , inc, dec, sub, add, xor clearing, xor general, and, push, pop. Will have its own function to generate and replace it with its alternate equivalent.

Note that the adding of junk code and alternate instructions are based on a probability, which is a const variable in the Morph_Executable_Controller.h, called INSERTION_PROBABILITY. This is set at 15 as it will not bloat up the amount of bytes. Hence, the morphed executable file size will not be significantly larger than its original executable file.