In the Compile_Code_Controller, we refracted the status variable to analysis_textbox_status, to be more clear. Also, we added a new function to enable the Compile_Menu_Item, when the user opens an existing file or creates a new file.
In the View_Compilation_Status_Controller, we refracted the function name of update_analysis_textbox to update_analysis_textbox_and_enable_morph. Inside this function, we also enabled the Morph_Button and Morph_Menu_Item when the user sucessfully compiles code.

In the Open_Existing_File_Controller and the Create_New_File_Controller, we called the set_enabled_compile_menu_item when the user opens an existing file or creates a new file.

In the Morph_Executable_Controller, we added a function to called update_analysis_textbox which updates the analysis textbox, using other supporting functions, which gets the existing content from the analysis textbox in the UI and appends the morphing status when the Morph_Button or Morph_Menu_Item is clicked. Also in the morph_exe_with_encryption_junk_alt_instructions function, we calculated the time taken when this function is run and it will be displayed into the analysis textbox through the update_analysis_textbox function