In the MainWindow, we refractored set_exe_file_path to set_original_exe_file_path, this is to provide more clarity to the function. In addition, we have added 2 labels, called Original_Executable_Label and Morph_Executable_Label, and 2 buttons. called Run_Original_Button and Run_Morphed_Button. Also, when compiling and morphing, we will store the full file path and the executable file names in variables, with the accessor and mutator functions for each.
In Compile_Code_Controller, we added afew functions to get the full file path and the executable name. Similarly, In the Morph_Executable_Controller, we added afew functions to get the full morphed file path and the morphed executable name.

In View_Compilation_Status_Controller, we created a new variable to store the compilation status, this is to make sure that upon successful compilation, it matches with the success message in the error.h ( this success message has been newly added ).

We created the Run_Executable_Controller class. This class will update the above mentioned labels with the respective executable names. Also, it is responsible for running the respective executables via command line upon clicking the above mentioned buttons in the MainWindow.