In the MainWindow, we added the View_Signatures_Button and the View_Sizes_Button.
In the Morph_Executable_Controller, in the error_warning_message_box, we changed ERROR_INVALID_EXECUTABLE to ERROR_FILE_NOT_FOUND. In addition, we added a function called add_conditional_jumps_instructions, which adds in conditional jumps in the payload. We reduced the probability, from 15 to 10 percent, to prevent bloating of the payload section.

We added View_Executable_Signatures_Controller and View_Executable_Sizes_Controller, they will get the executable file MD5 signatures and executable file sizes corresponding to the classes.

In the File_Reader, we added a method to get the executable file size.