In the Morph_Executable_Controller, we added a function called concat_disassembly, which concatinates the disassembly vector into a single QString. This disassembly vector will contain the disassembly of the payload before and after adding junk and alternate instructions.
We created a View_Disassembly_Controller, the objective of this class is to save and spawn a text file, which contains the disassembly mentioned above. We have a function to enable its respective button when morphing is successful.

In the View_Executable_Sizes_Controller and the View_Executable_Signature_Controller, we changed the format_text function to include original and morphed for more clarity when it is displayed. These classes also have functions to enable their respective buttons when compilation is successful.