Added more payloads in the payloads.h. The respective radio buttons were made and the payloads were added in the Morph_Executable_Controller's morphing functions.
In the Morph_Executable_Controller, while concatinating the dissassembly, we added a time stamp for clarification for when it was morphed.

Changed the way we get the information from the radio button using QButtonGroup to get the ID, instead of matching object name and is checked in the old method. Also upon clicking the respective radio buttons, the Choose_Payload_Controller will display the respective information of the payload in the newly added Payload_Info_Textbox.

We also revamped the UI to make it look more stylized and clean. In addition, we changed the format any text inside the analysis text box. Also, we increased the overall font size of the GUI to make it more easier to read.

We also added high dpi scaling, in hopes that the resolution will be standardized when this polymorphic engine runs on different machines.